[streamrestore] Add stream volume transform feature.
New option in configuration, with prefix transform, allows one to track
one stream volume, and whenever the volume changes the changed volume is
set to other stream, with possibility to restrict the volume range to
different range from the source stream.

For example:
transform.dtmf = phone;dtmf;40;80

Here whenever stream with stream restore id phone volume changes, the
volume is then translated to the range of 40..80 and set to other
stream with stream restore id dtmf.