[haptic] Divide all haptic types to two classes. Fixes JB#38974
When new haptic types were introduced the new types were no longer
properly handled by the haptic feature of ngfd. Split all known haptic
types to two classes, HAPTIC_CLASS_TOUCH and HAPTIC_CLASS_ALARM, and
filter the events based on the class and profile settings.

Currently the known types and classes are defined in the ngfd source,
but if more flexible configuration is needed in the future the
definition needs to be somehow modified to be read from ini files.