[resouce] Update resource filtering to type based.
Previously sink name was used to determine when filtering
sinks out, but it makes more sense to have higher level
category defining what to filter. So use sink interface
type to decide when to filter sinks for a request.

Also, all sink types are enabled by default, and all types
that need to be filtered need to be marked for filtering
individually (previously setting only one or more flag as
enabled would set all remaining flags to disabled).