[canberra] Update request proplist after uploading. JB#40097
When sample is first uploaded to PulseAudio it contains all the
properties from the moment of upload. When the sample is later played
and the new play request contains different values for properties, those
are overridden from the uploaded sample proplist, but already existing
properties cannot be dropped. Because of this we need to upload the
sample with minimum properties so that we have better control when
actually playing the sample.