[tonegen] Refactor tone object create / destroy functions. JB#63153
There have been ngfd crashes with tone_destroy() included in
backtraces and code looks suspicious and hard to follow.

After closer look, the code does not contain any obvious faults.
Try to make it a bit more readable, or at least add comments
for the benefit of the next reader.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>