[ngfd] Lower warning for no sinks handling feedback to debug level. JB#63363
Commit e1ec7892b8e enabled warning logging by default.
With e.g. haptic feedbacks of for touch screen events, it meant having
these in journal:
WARNING: core: no sinks that can handle the request 'feedback_press_strong'

Which sounds scarier than it is. In this case n_haptic_can_handle()
returns false based on configuration. Might be nice differentiating
better between anything able to handle and anything willing to handle
right now, but that would be bigger changes. So now just adjusting
logging level and message a bit now so that it's not confused as ngfd
not being capable for handling the event.